Add serialization tests for all classes on the wire
